fix to ensure memory constraint on output
Previously I had allowed the main function merge-to-outfile to build
up the rows of the output csv in memory before writing at once.
Now the function takes an output writer and writes output to that one
row at a time.  Terrible performance, but guarantees memory constraint.
This commit keeps old version of the function as well.

Add a function to read all input files into memory at once and return
their line count for testing.  Give this an option in -main.